requirements.txt
# Core AI/ML Framework
langchain>=0.1.0
langgraph>=0.0.1
langchain-groq>=0.0.1
pydantic>=2.0.0
email-validator>=2.0.0

# Web Framework
streamlit>=1.28.0

# Document Processing
pymupdf>=1.23.0
python-docx>=0.8.11
reportlab>=4.0.0
json-repair>=0.19.0

# Semantic Search & Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Environment & Configuration
python-dotenv>=1.0.0